refactor: move some methods to class Storage
